requirements.txt
pydantic==2.11.1
python-telegram-bot==20.7
ccxt==4.3.63
pandas==2.2.2
numpy==1.26.4
aiohttp==3.9.5
python-dotenv==1.0.1
PyYAML==6.0.2
